Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/parsing/transition-duration-computed-expected.txt


PASS Property transition-duration value '500ms'
PASS Property transition-duration value 'calc(2 * 3s)'
FAIL Property transition-duration value '20s, 10s' assert_equals: expected "20s, 10s" but got "10s"
